+++ gadgets/mri_core/CropAndCombineGadget.cpp
@@ -11,7 +11,7 @@ process( GadgetContainerMessage<ISMRMRD::ImageHeader>* m1,
   GadgetContainerMessage< hoNDArray< std::complex<float> > >* m3 = 
     new GadgetContainerMessage< hoNDArray< std::complex<float> > >();
 
-  std::vector<unsigned int> new_dimensions(3);
+  std::vector<size_t> new_dimensions(3);
   new_dimensions[0] = m2->getObjectPtr()->get_size(0)>>1;
   new_dimensions[1] = m2->getObjectPtr()->get_size(1);
   new_dimensions[2] = m2->getObjectPtr()->get_size(2);
