+++ test/hoNDArray_blas_test.cpp
@@ -9,12 +9,12 @@ using testing::Types;
 template <typename T> class hoNDArray_blas_Real : public ::testing::Test {
 protected:
   virtual void SetUp() {
-    unsigned int vdims[] = {37, 49, 23, 19}; //Using prime numbers for setup because they are messy
-    dims = std::vector<unsigned int>(vdims,vdims+sizeof(vdims)/sizeof(unsigned int));
+    size_t vdims[] = {37, 49, 23, 19}; //Using prime numbers for setup because they are messy
+    dims = std::vector<size_t>(vdims,vdims+sizeof(vdims)/sizeof(size_t));
     Array = hoNDArray<T>(&dims);
     Array2 = hoNDArray<T>(&dims);
   }
-  std::vector<unsigned int> dims;
+  std::vector<size_t> dims;
   hoNDArray<T> Array;
   hoNDArray<T> Array2;
 };
@@ -72,12 +72,12 @@ TYPED_TEST(hoNDArray_blas_Real,amaxTest){
 template <typename T> class hoNDArray_blas_Cplx : public ::testing::Test {
 protected:
   virtual void SetUp() {
-    unsigned int vdims[] = {37, 49}; //Using prime numbers for setup because they are messy
-    dims = std::vector<unsigned int>(vdims,vdims+sizeof(vdims)/sizeof(unsigned int));
+    size_t vdims[] = {37, 49}; //Using prime numbers for setup because they are messy
+    dims = std::vector<size_t>(vdims,vdims+sizeof(vdims)/sizeof(size_t));
     Array = hoNDArray<T>(&dims);
     Array2 = hoNDArray<T>(&dims);
   }
-  std::vector<unsigned int> dims;
+  std::vector<size_t> dims;
   hoNDArray<T> Array;
   hoNDArray<T> Array2;
 };
